+++ lib/polar.dart
@@ -7,9 +7,7 @@ export 'src/model/device_streaming_feature.dart' show DeviceStreamingFeature;
 export 'src/model/ohr_data_type.dart' show OhrDataType;
 export 'src/model/polar_device_info.dart';
 export 'src/model/polar_hr_data.dart';
-export 'src/model/polar_ohr_ppi_sample.dart';
 export 'src/model/polar_recording.dart';
 export 'src/model/polar_sensor_setting.dart'
     show PolarSensorSetting, PolarSettingType;
 export 'src/model/polar_streaming.dart';
-export 'src/model/xyz.dart';