+++ lib/src/model/converters.dart
@@ -1,0 +1,29 @@
+import 'dart:io';
+
+import 'package:json_annotation/json_annotation.dart';
+
+/// Converts platform values to booleans
+/// - The iOS SDK uses `0` and `1` for booleans
+/// - The Android SDK uses `true` and `false` for booleans
+class PlatformBooleanConverter extends JsonConverter<bool, dynamic> {
+  /// Constructor
+  const PlatformBooleanConverter();
+
+  @override
+  bool fromJson(dynamic json) {
+    if (Platform.isIOS) {
+      return json != 0;
+    } else {
+      return json;
+    }
+  }
+
+  @override
+  dynamic toJson(bool object) {
+    if (Platform.isIOS) {
+      return object ? 1 : 0;
+    } else {
+      return object;
+    }
+  }
+}