+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.0.0-beta0
+- Updates Polar SDK to 4.0.0-beta0
+- BREAKING: Moves `timeStamp` field from streaming data into individual samples
+- BREAKING: Renames PolarOhrPpiSample to PolarPpiSample
+- BREAKING: Streaming data no longer contains `identifier` fields
+
 ## 3.7.0
 - Fixes issue with PPI streaming